feat: enum
