+++ App/UI/GreenCertificate/ModalDetail/GreenCertificateVaccineDetail/GreenCertificateVaccineDetailVC.swift
